+++ CHANGELOG.md
@@ -4,6 +4,38 @@
 
 ### Added
 
+- Added service `Services\Sale\Shipment\Service\Shipment` with support methods,
+  see [sale.shipment.* methods](https://github.com/bitrix24/b24phpsdk/issues/250):
+    - `add` adds a shipment
+    - `update` updates the fields of a shipment
+    - `get` returns a shipment by its identifier
+    - `list` returns a list of shipments
+    - `delete` deletes a shipment
+    - `getFields` returns the fields and settings for shipments
+- Added service `Services\Sale\ShipmentProperty\Service\ShipmentProperty` with support methods,
+  see [sale.shipmentproperty.* methods](https://github.com/bitrix24/b24phpsdk/issues/250):
+    - `add` adds a shipment property
+    - `update` updates the fields of a shipment property
+    - `get` returns a shipment property by its identifier
+    - `list` returns a list of shipment properties
+    - `delete` deletes a shipment property
+    - `getFieldsByType` returns the fields and settings for shipment properties by type
+- Added service `Services\Sale\ShipmentPropertyValue\Service\ShipmentPropertyValue` with support methods,
+  see [sale.shipmentpropertyvalue.* methods](https://github.com/bitrix24/b24phpsdk/issues/250):
+    - `modify` updates shipment property values for a shipment
+    - `get` returns a shipment property value by its identifier
+    - `list` returns a list of shipment property values
+    - `delete` deletes a shipment property value
+    - `getFields` returns the fields and settings for shipment property values
+
+- Added service `Services\Sale\ShipmentItem\Service\ShipmentItem` with support methods,
+  see [sale.shipmentitem.* methods](https://github.com/bitrix24/b24phpsdk/issues/250):
+    - `add` adds a new shipment item
+    - `update` updates the fields of a shipment item
+    - `get` returns a shipment item by its identifier
+    - `list` returns a list of shipment items
+    - `delete` deletes a shipment item
+    - `getFields` returns the fields and settings for shipment items
 - Added service `Services\Sale\Payment\Service\Payment` with support methods,
   see [sale.payment.* methods](https://github.com/bitrix24/b24phpsdk/issues/248):
     - `add` adds a payment